requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
confluent-kafka==2.5.0
requests==2.32.3
sentence-transformers==3.0.1
numpy==1.26.4
google-generativeai==0.8.1
python-dotenv==1.0.1
